requirements.txt
ansiblemetrics>=0.3.6
django~=3.1.1
django-rest-swagger~=2.2.0
djangorestframework~=3.12.1
repositories-collector~=0.0.2
repository-miner>=0.2.7
repository-scorer>=0.0.5
pandas>=1.1.2
pydriller>=1.15.2
pygithub>=1.53
pymongo~=3.11.0
python-dotenv~=0.14.0
scikit-learn>=0.23.2
imbalanced-learn>=0.7.0
jsonpickle>=1.4.1